handlers.go: allow overriding WireGuard endpoint
When running on platforms like Kubernetes, the domain name used for
handling HTTP may not be the name of the host running WireGuard. For
example, HTTP ingress is load balanced across all Kubernetes nodes in
the cluster but WireGuard is only run on a single node in the cluster.
This commit enables setting the WireGuard endpoint independently of the
HTTP hostname. If `SUBSPACE_ENDPOINT_HOST` is unset, it defaults to the
value of `SUBSPACE_HTTP_HOST` so that the old behavior is kept.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>